Only request zoom library if it is not disabled.
There is already a if condition in `footer.html` that only runs the JS
from the zoom library if it is enabled per-site.
This commit takes the same if condition and applies it in `head.html` as
well to save a unnecessary web request to `zoom.min.js` if that is the case.